[un-/viewModel] mess with type checking
viewModel.merge is tricky to type. The solution here is to require that createViewModel<T>() be given the final type we want.

The vm initializers are all Partials, and the merge initializers are Partials of both T and the Source type S.

The previous system of building up the type of the merged VM from T & S required very careful typing of the initializers, using Pick<T, "" | "" >